Subclassing CLoudAuxWatcher for IAMUser.
