+++ security_monkey/watchers/cloudaux_s3.py
@@ -15,10 +15,10 @@ class S3(CloudAuxWatcher):
         buckets = list_buckets(**kwargs)['Buckets']
         return [bucket['Name'] for bucket in buckets if not self.check_ignore_list(bucket['Name'])]
 
-    def get_method(self, item, **kwargs):
-        bucket = get_bucket(item, **kwargs)
+    def get_method(self, item_name, **kwargs):
+        bucket = get_bucket(item_name, **kwargs)
         
         if bucket and bucket.get("Error"):
-            raise SecurityMonkeyException("S3 Bucket: {} fetching error: {}".format(item, bucket["Error"]))
+            raise SecurityMonkeyException("S3 Bucket: {} fetching error: {}".format(item_name, bucket["Error"]))
         
-        return bucket
+        return bucket, item_name
